trying to fix device issues

+++ diffusion.py
@@ -14,7 +14,7 @@ def get_index_from_list(vals, t, x_shape):
     while considering the batch dimension.
     """
     batch_size = t.shape[0]
-    out = vals.gather(-1, t)
+    out = vals.gather(-1, t.long())
     return out.reshape(batch_size, *((1,) * (len(x_shape) - 1)))
 
 def forward_diffusion_sample(x_0, t, T=300):
